+++ Frontend/static-node-app/package.json
@@ -16,7 +16,8 @@
     "react-dom": "^18.2.0",
     "react-icons": "^5.1.0",
     "react-responsive": "^10.0.0",
-    "react-router-dom": "^6.23.0"
+    "react-router-dom": "^6.23.0",
+    "react-type-animation": "^3.2.0"
   },
   "devDependencies": {
     "@types/react": "^18.2.66",
