+++ Frontend/static-node-app/src/App.css
@@ -32,11 +32,6 @@
   }
 }
 
-h1 {
-  font-size: 3.2em;
-  line-height: 1.1;
-}
-
 .appBody{
   display: flex;
   flex-direction: column;
