+++ Frontend/static-node-app/src/components/Footer.jsx
@@ -1,4 +1,3 @@
-//import React from 'react';
 import "../styles/Footer.css";
 function Footer (){
     return (
