--- Frontend/static-node-app/src/components/Home.jsx
@@ -1,12 +1,0 @@
-import React from "react";
-
-const Home = () => {
-    return (
-        <>
-         <p>This component will soon render the homepage</p>
-        </>
-    );
-   };
-   
-export default Home;
-
