+++ Frontend/static-node-app/src/index.css
@@ -8,8 +8,7 @@
   font-weight: 400;
 
   color-scheme: light dark;
-  color: #13361e;
-  background-color: #d7e5d6;
+  background-color: #e4e7d2;
 
   font-synthesis: none;
   text-rendering: optimizeLegibility;
