+++ Frontend/static-node-app/src/styles/Footer.css
@@ -2,7 +2,7 @@
     width: 100vw;
     display: flex;
     justify-content: center;
-    background-color: #d7e5d6;
+    background-color: #e4e7d2;
     margin-top: auto;
     bottom: 0;
 }
