+++ Frontend/static-node-app/src/styles/HomePage.css
@@ -1,0 +1,12 @@
+.headerDiv{
+    display: flex;
+    justify-content: center;
+    background-color: rgb(1,1,1,0.05);
+    height: 100px;
+    padding: 20px;
+    margin-top: 100px;
+    margin-right: 15%;
+    margin-left: 15%;
+    border-radius: 15px;
+    box-shadow: 10px 5px 5px #bbc591;
+}
