+++ Frontend/static-node-app/src/styles/LoginPage.css
@@ -1,0 +1,10 @@
+.loginDiv{
+    display: flex;
+    justify-content: center;
+    background-color: rgb(1,1,1,0.05);
+    box-shadow: 10px 5px 5px #bbc591;
+   margin-top: 100px;
+    height: 600px;
+    width: 500px;
+    border-radius: 30px;
+}
